news_chatbot_agent/pyproject.toml
[project]
name = "news-chatbot-agent"
version = "0.1.0"
description = "뉴스 기반 자연어 챗봇 에이전트 - 교육용"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-upstage>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.0",
    "requests>=2.31.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ipython>=8.0.0",
]
